+++ pkg/decrypt/vault_decrypter.go
@@ -4,20 +4,22 @@ import (
 	"context"
 	"encoding/base64"
 
+	"github.com/giantswarm/konfigure/internal/vaultclient"
+
 	"github.com/giantswarm/microerror"
-	vaultapi "github.com/hashicorp/vault/api"
+	"github.com/giantswarm/valuemodifier/vault/decrypt"
 )
 
 const (
-	keyring = "/v1/transit/decrypt/config"
+	key = "config"
 )
 
 type VaultDecrypterConfig struct {
-	VaultClient *vaultapi.Client
+	VaultClient *vaultclient.WrappedVaultClient
 }
 
 type VaultDecrypter struct {
-	vaultClient *vaultapi.Client
+	vaultClient *vaultclient.WrappedVaultClient
 }
 
 var _ Decrypter = &VaultDecrypter{}
@@ -35,51 +37,26 @@ func NewVaultDecrypter(config VaultDecrypterConfig) (*VaultDecrypter, error) {
 }
 
 func (d *VaultDecrypter) Decrypt(ctx context.Context, ciphertext []byte) ([]byte, error) {
-	req := struct {
-		Ciphertext string `json:"ciphertext"`
-	}{
-		Ciphertext: string(ciphertext),
+	if err := d.vaultClient.ConfigurationValidator(); err != nil {
+		return nil, microerror.Mask(err)
 	}
 
-	resp := struct {
-		Data struct {
-			Plaintext string `json:"plaintext"`
-		} `json:"data"`
-	}{}
+	service, err := decrypt.New(decrypt.Config{VaultClient: d.vaultClient.Wrapped, Key: key})
 
-	err := d.vaultRequest(ctx, keyring, req, &resp)
 	if err != nil {
 		return nil, microerror.Mask(err)
 	}
 
-	decoded, err := base64.StdEncoding.DecodeString(resp.Data.Plaintext)
+	plainText, err := service.Decrypt(ciphertext)
+
+	if err != nil {
+		return nil, microerror.Mask(err)
+	}
+
+	decoded, err := base64.StdEncoding.DecodeString(plainText)
 	if err != nil {
 		return nil, microerror.Mask(err)
 	}
 
 	return decoded, nil
-}
-
-func (d *VaultDecrypter) vaultRequest(ctx context.Context, endpoint string, req, resp interface{}) error {
-	httpReq := d.vaultClient.NewRequest("POST", endpoint)
-	err := httpReq.SetJSONBody(req)
-	if err != nil {
-		return microerror.Mask(err)
-	}
-
-	httpResp, err := d.vaultClient.RawRequest(httpReq)
-	if err != nil {
-		return microerror.Mask(err)
-	}
-
-	if httpResp.StatusCode != 200 {
-		return microerror.Maskf(executionFailedError, "expected status code = 200, got %d", httpResp.StatusCode)
-	}
-
-	err = httpResp.DecodeJSON(resp)
-	if err != nil {
-		return microerror.Mask(err)
-	}
-
-	return nil
 }