+++ CHANGELOG.md
@@ -7,6 +7,32 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.11.0] - 2022-08-03
+
+### Changed
+
+- The Vault client configuration is validated when `konfigure` actually tries to decrypt something with it instead at initialisation time, so e.g. `VAULT_*` environment variables can be omitted now when Vault is not used.
+
+## [0.10.0] - 2022-07-21
+
+## [0.9.0] - 2022-06-27
+
+### Fixed
+
+- `konfigure lint` now respects template escape markers.
+
+## [0.8.0] - 2022-05-25
+
+### Changed
+
+- Update dependencies to support Flux v0.30.2.
+
+## [0.7.0] - 2022-05-10
+
+### Added
+
+- Support SOPS with GnuPGP and AGE encryption.
+
 ## [0.6.0] - 2022-04-14
 
 ### Added
@@ -155,7 +181,12 @@ blocked by app-admission-controller.
 - Start of a new history. Previous commits have been imported from https://github.com/giantswarm/config-controller.
 - Use local filesystem instead of GitHub as configuration source.
 
-[Unreleased]: https://github.com/giantswarm/konfigure/compare/v0.6.0...HEAD
+[Unreleased]: https://github.com/giantswarm/konfigure/compare/v0.11.0...HEAD
+[0.11.0]: https://github.com/giantswarm/konfigure/compare/v0.10.0...v0.11.0
+[0.10.0]: https://github.com/giantswarm/konfigure/compare/v0.9.0...v0.10.0
+[0.9.0]: https://github.com/giantswarm/konfigure/compare/v0.8.0...v0.9.0
+[0.8.0]: https://github.com/giantswarm/konfigure/compare/v0.7.0...v0.8.0
+[0.7.0]: https://github.com/giantswarm/konfigure/compare/v0.6.0...v0.7.0
 [0.6.0]: https://github.com/giantswarm/konfigure/compare/v0.5.6...v0.6.0
 [0.5.6]: https://github.com/giantswarm/konfigure/compare/v0.5.5...v0.5.6
 [0.5.5]: https://github.com/giantswarm/konfigure/compare/v0.5.4...v0.5.5